+++ leaf-fe-vue/Dockerfile
@@ -1,0 +1,8 @@
+FROM node:lts-alpine
+RUN npm install -g http-server
+WORKDIR /app
+COPY package*.json ./
+RUN npm install
+COPY . . 
+EXPOSE 8080
+CMD ["npm", "run", "serve"]
