+++ leaf-fe-vue/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "leaf-fe",
+  "name": "leaf_fe",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -7,9 +7,16 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
+    "@apollo/client": "^3.5.5",
+    "@vue/apollo-composable": "^4.0.0-alpha.16",
+    "@vue/apollo-option": "^4.0.0-alpha.16",
+    "apollo-link-context": "^1.0.20",
     "core-js": "^3.6.5",
+    "graphql": "^15.8.0",
+    "graphql-tag": "^2.12.6",
     "vue": "^3.0.0",
-    "vue-router": "^4.0.0-0"
+    "vue-router": "^4.0.0-0",
+    "vuex": "^3.6.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
