+++ leaf-fe-vue/src/App.vue
@@ -1,643 +1,66 @@
 <template>
-<div class="container">
+  <div id="app">
 
-        <div class="forms-container">
+    <!-- <div>
+      <button v-if="is_auth" v-on:click="logout">
+          <span>Cerrar sesión</span>
+      </button>
+    </div> -->
 
-            <div class="signin-signup">
+    <router-view v-on:completedLogin="completedLogin"></router-view>
 
-                <form v-on:submit.prevent="processLogIn" class="sign-in-form">
-
-                    <h2 class="title">Iniciar Sesión</h2>
-                    <div class="input-field">
-                        <i class="fas fa-envelope"></i>
-                        <input v-model="credentials.username" type="text" placeholder="Username">
-                    </div>
-                    <div class="input-field">
-                        <i class="fas fa-lock"></i>
-                        <input v-model="credentials.password" type="password" placeholder="Contraseña">
-                    </div>
-
-                    <p v-if="show_error" class="error">Usuario o contraseña incorrectos</p> 
-
-                    <input v-bind:class="{'disabled': is_loading}" type="submit" value="Ingresar" class="btn solid">
-                    <p class="social-text">O ingresa a nuestras redes sociales</p>
-
-                    <div class="social-media">
-
-                        <a href="#" class="social-icon"><i class="fab fa-facebook-f"></i></a>
-                        <a href="#" class="social-icon"><i class="fab fa-twitter"></i></a>
-                        <a href="#" class="social-icon"><i class="fab fa-instagram"></i></a>
-                        <a href="#" class="social-icon"><i class="fab fa-linkedin"></i></a>
-
-                    </div>
-
-                </form>
-
-                <form action="#" class="sign-up-form">
-
-                    <h2 class="title">Registrarse</h2>
-
-                    <div class="input-field">
-                        <i class="fas fa-user"></i>
-                        <input type="text" placeholder="Nombre Completo"> <!-- Razón Social -->
-                    </div>
-                    <div class="input-field">
-                        <i class="fas fa-envelope"></i>
-                        <input type="email" placeholder="Correo">
-                    </div>
-                    <div class="input-field">
-                        <i class="fas fa-map-marker-alt"></i>
-                        <input type="text" placeholder="Dirección">
-                    </div>
-                    <div class="input-field">
-                        <i class="fas fa-lock"></i>
-                        <input type="password" placeholder="Contraseña">
-                    </div>
-                    <input type="submit" value="Registrarse" class="btn">
-                    <p class="social-text">O ingresa a nuestras redes sociales</p>
-
-                    <div class="social-media">
-
-                        <a href="#" class="social-icon"><i class="fab fa-facebook-f"></i></a>
-                        <a href="#" class="social-icon"><i class="fab fa-twitter"></i></a>
-                        <a href="#" class="social-icon"><i class="fab fa-instagram"></i></a>
-                        <a href="#" class="social-icon"><i class="fab fa-linkedin"></i></a>
-
-                    </div>
-
-
-                </form>
-
-            </div>
-
-        </div>
-
-        <div class="panels-container">
-
-            <div class="panel left-panel">
-
-                <div class="content">
-
-                    <h3>¿Eres nuevo por aquí?</h3>
-                    <p>Lorem ipsum dolor sit, amet consectetur adipisicing elit. Iste sapiente odio atque quas quos. </p>
-                    <button v-on:click="signUpMode" class="btn transparent" id="sign-up-btn">
-                        Registrarse
-                    </button>
-
-                </div>
-                <img src="..\src\assets\log_user.svg" alt="Login User" class="image">
-
-            </div>
-            <div class="panel right-panel">
-
-                <div class="content">
-
-                    <h3>¿Ya tienes una cuenta?</h3>
-                    <p>Lorem ipsum dolor sit, amet consectetur adipisicing elit. Reiciendis ex odit quisquam adipisci! Dicta, magnam?</p>
-                    <button class="btn transparent" id="sign-in-btn">Ingresar</button>
-
-                </div>
-                <img src="..\src\assets\registerUser.svg" alt="Register User" class="image">
-
-            </div>
-
-        </div>
-
-    </div>
+  </div>
 </template>
-<script>
-import gql from "graphql-tag";
 
+<script>
 export default {
     name: "App", // Nombre del componente
-
     data: function() {
-      return { 
-          is_auth: false,
-          show_error: false,
-          is_loading: false,
-          credentials: {
-              username: "",
-              password: ""
-        }
-       }
+      return {
+        is_auth: false
+      }
     }, // Todas las variables de este componentes
-
     methods: {
       verifyAuth(){
         this.is_auth = localStorage.getItem("is_auth") || false;
+        console.log(this.is_auth);
         if(this.is_auth){
-          this.loadCarrito();
+          this.loadIndex();
         }else{
-          //this.loadLogIn();
+          this.loadLogIn();
         }
       },  
+      loadLogIn(){
+        this.$router.push({name: "logIn"});
+      },
+      loadSignUp(){
+        this.$router.push({name: "signUp"});
+      },
+      loadCarrito(){
+        this.$router.push({name: "carrito"});
+      },
+      loadIndex(){
+        this.$router.push({name: "index"});
+      },
       completedLogin(data){
         this.is_auth = true;
         localStorage.setItem("is_auth", true);
         localStorage.setItem("token", data.token);
         localStorage.setItem("username", data.username);
-        console.log(`token: ${localStorage.getItem("token")}`);
         this.verifyAuth()
       },
-      signUpMode(){
-        this.classList.add("sign-up-mode");
-      },
-      loadCarrito(){
-        console.log("Estoy en LoadCarrito");
-        this.$router.push({name: 'carrito'});
-      },
-      processLogIn: async function (){
-            this.is_loading = true;
-            await this.$apollo.mutate({
-                mutation: gql`
-                    mutation Login($credentials: LoginInput!) {
-                        login(credentials: $credentials) {
-                            key
-                        }
-                    }
-                `,
-                variables: {
-                    credentials: this.credentials
-                }
-            })
-            .then((result) => {
-                console.log("FUNCIONÓOOO")
-                this.is_loading = false;
-                this.show_error = false;
-                let data = {
-                    username: this.credentials.username,
-                    token: result.data.login.key
-                }
-                this.completedLogin(data);
-            })
-            .catch((error)=>{
-                this.show_error = true;
-                console.log("DIO ERROR :c")
-                console.log(error)
-                this.is_loading = false;
-            })
-        }
+      logout(){
+        this.is_auth = false;
+        localStorage.clear();
+        this.verifyAuth()
+      }
     }, // Todas las funciones que usa este componente
-    
     created: function () {
-
+      this.loadIndex();
     } // Eventos: lo que pasa cuando el componente se inicia
 };
 </script>
 
 <style>
-*{
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    font-family: sans-serif;
-}
-.container {
-  position: relative;
-  width: 100%;
-  background-color: #fff;
-  min-height: 100vh;
-  overflow: hidden;
-}
 
-.forms-container {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left: 0;
-  }
-  
-
-.signin-signup {
-
-    position: absolute;
-    top: 50%; left: 75%;
-    transform: translate(-50%, -50%);
-    width: 50%;
-    transition: 1s 0.7s ease-in-out;
-    display: grid;
-    grid-template-columns: 1fr;
-    z-index: 5;
-
-}
-
-form {
-
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    padding: 0 5rem;
-    transition: all 1s 0.7s;
-    overflow: hidden;
-    grid-column: 1 / 2;
-    grid-row: 1 / 2;
-
-}
-
-form.sign-up-form {
-
-    opacity: 0;
-    z-index: 1;
-
-}
-
-form.sign-in-form {
-
-    z-index: 2;
-
-}
-
-.title {
-
-    font-size: 2.2rem;
-    color: #444;
-    margin-bottom: 10px;
-
-}
-
-.input-field {
-
-    position: relative;
-    max-width: 380px;
-    width: 100%;
-    background-color: #f0f0f0;
-    margin: 10px 0;
-    height: 55px;
-    border-radius: 55px;
-    grid-template-columns: 15% 85%;
-    padding: 0 .4rem;
-
-}
-
-.input-field i {
-
-    text-align: center;
-    line-height: 55px;
-    color: #acacac;
-    transition: 0.5s;
-    font-size: 1.1rem;
-    margin: 0 10px;
-
-}
-
-.input-field input {
-
-    background: none;
-    border: none; outline: none;
-    line-height: 1;
-    font-weight: 600;
-    font-size: 1.1rem;
-    color: #333;
-
-}
-
-.input-field input::placeholder {
-
-    color: #aaa;
-    font-weight: 500;
-
-
-}
-
-.social-text {
-
-    padding: 0.7rem 0;
-    font-size: 1rem;
-
-}
-
-.social-media {
-
-    display: flex;
-    justify-self: center;
-    transition: 0.3s;
-    transform: translateY(-0px);
-
-}
-
-.social-icon {
-
-    height: 46px; width: 46px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin: 0 .45rem;
-    color: #333;
-    border: 1px solid #333;
-    border-radius: 50%;
-    text-decoration: none;
-    font-size: 1.1rem;
-    transition: 0.3s;
-
-}
-
-.social-icon:hover {
-
-    color: #fff;
-    background-color: #333;    
-
-}
-
-.btn {
-
-    width: 150px;
-    background-color: #3e6553;
-    color: #fff;
-    border: none; outline: none;
-    height: 49px;
-    border-radius: 49px;
-    font-size: 1rem;
-    text-transform: uppercase;
-    font-weight: 600;
-    margin: 10px 0;
-    cursor: pointer;
-    transition: 0.5s;
-
-}
-
-.btn:hover {
-
-    background-color: #365949;
-
-}
-
-.panels-container {
-
-    position: absolute;
-    top: 0; left: 0;
-    width: 100%; height: 100%;
-    background-color: #fff;
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-
-}
-
-.container:before {
-
-    content: '';
-    position: absolute;
-    top: -10px; right: 48%;
-    width: 2000px; height: 2000px;
-    transform: translateY(-50%);
-    background-image: linear-gradient(-45deg, #365949 0%, #93bdaa 100%);
-    transition: 1.8s ease-in-out;
-    border-radius: 50%;
-    z-index: 6;
-
-}
-
-.image {
-
-    width: 100%;
-    transition: transform 0.9s ease-in-out;
-    transition-duration: .6s;
-
-}
-
-.panel {
-
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    justify-content: space-between;
-    text-align: center;
-    z-index: 6;
-
-}
-
-.left-panel {
-
-    pointer-events: all;
-    padding: 3rem 17% 2rem 12%;
-
-}
-.right-panel {
-
-    pointer-events: none;
-    padding: 3rem 17% 2rem 17%;
-
-}
-
-.panel .content {
-
-    color: #fff;
-    transition: transform 0.9s ease-in-out;
-    transition-delay: .6s;
-
-}
-
-.panel h3 {
-
-    font-weight: 600;
-    line-height: 1;
-    font-size: 1.5rem;
-
-}
-
-.panel p {
-
-    font-size: .95rem;
-    padding: 0.7rem 0;
-
-}
-
-.btn.transparent {
-
-    margin: 0;
-    margin-top: .2rem;
-    background: none;
-    border: 2px solid #fff;
-    width: 130px; height: 41px;
-    font-weight: 600;
-    font-size: .8rem;
-
-}
-
-.error{
-    color: red;
-    font-size: 15px;
-}
-
-.disabled{
-    pointer-events: none;
-    opacity: 0.6;
-}
-
-.right-panel .image,
-.right-panel .content {
-
-    transform: translateX(800px);
-
-}
-
-/* ANIMACIONES */
-
-.container.sign-up-mode:before {
-    transform: translate(100%, -50%);
-    right: 52%;
-  }
-  
-  .container.sign-up-mode .left-panel .image,
-  .container.sign-up-mode .left-panel .content {
-    transform: translateX(-800px);
-  }
-  
-  .container.sign-up-mode .signin-signup {
-    left: 25%;
-  }
-  
-  .container.sign-up-mode form.sign-up-form {
-    opacity: 1;
-    z-index: 2;
-  }
-  
-  .container.sign-up-mode form.sign-in-form {
-    opacity: 0;
-    z-index: 1;
-  }
-  
-  .container.sign-up-mode .right-panel .image,
-  .container.sign-up-mode .right-panel .content {
-    transform: translateX(0%);
-  }
-  
-  .container.sign-up-mode .left-panel {
-    pointer-events: none;
-  }
-  
-  .container.sign-up-mode .right-panel {
-    pointer-events: all;
-  }
-
-  @media (max-width: 570px) {
-    form {
-      padding: 0 1.5rem;
-    }
-  
-    .image {
-      display: none;
-    }
-    .panel .content {
-      padding: 0.5rem 1rem;
-    }
-    .container {
-      padding: 1.5rem;
-    }
-  
-    .container:before {
-      bottom: 72%;
-      left: 50%;
-    }
-  
-    .container.sign-up-mode:before {
-      bottom: 28%;
-      left: 50%;
-    }
-  }
-
-  @media (max-width: 870px) {
-    .container {
-      min-height: 800px;
-      height: 100vh;
-    }
-    .signin-signup {
-      width: 100%;
-      top: 95%;
-      transform: translate(-50%, -100%);
-      transition: 1s 0.8s ease-in-out;
-    }
-  
-    .signin-signup,
-    .container.sign-up-mode .signin-signup {
-      left: 50%;
-    }
-  
-    .panels-container {
-      grid-template-columns: 1fr;
-      grid-template-rows: 1fr 2fr 1fr;
-    }
-  
-    .panel {
-      flex-direction: row;
-      justify-content: space-around;
-      align-items: center;
-      padding: 2.5rem 8%;
-      grid-column: 1 / 2;
-    }
-  
-    .right-panel {
-      grid-row: 3 / 4;
-    }
-  
-    .left-panel {
-      grid-row: 1 / 2;
-    }
-  
-    .image {
-      width: 200px;
-      transition: transform 0.9s ease-in-out;
-      transition-delay: 0.6s;
-    }
-  
-    .panel .content {
-      padding-right: 15%;
-      transition: transform 0.9s ease-in-out;
-      transition-delay: 0.8s;
-    }
-  
-    .panel h3 {
-      font-size: 1.2rem;
-    }
-  
-    .panel p {
-      font-size: 0.7rem;
-      padding: 0.5rem 0;
-    }
-  
-    .btn.transparent {
-      width: 110px;
-      height: 35px;
-      font-size: 0.7rem;
-    }
-  
-    .container:before {
-      width: 1500px;
-      height: 1500px;
-      transform: translateX(-50%);
-      left: 30%;
-      bottom: 68%;
-      right: initial;
-      top: initial;
-      transition: 2s ease-in-out;
-    }
-  
-    .container.sign-up-mode:before {
-      transform: translate(-50%, 100%);
-      bottom: 32%;
-      right: initial;
-    }
-  
-    .container.sign-up-mode .left-panel .image,
-    .container.sign-up-mode .left-panel .content {
-      transform: translateY(-300px);
-    }
-  
-    .container.sign-up-mode .right-panel .image,
-    .container.sign-up-mode .right-panel .content {
-      transform: translateY(0px);
-    }
-  
-    .right-panel .image,
-    .right-panel .content {
-      transform: translateY(300px);
-    }
-  
-    .container.sign-up-mode .signin-signup {
-      top: 5%;
-      transform: translate(-50%, 0);
-    }
-  }
 </style>
