+++ leaf-fe-vue/src/components/signup.vue
@@ -1,0 +1,637 @@
+<template>
+<div class="container">
+
+        <div class="forms-container">
+
+            <div class="signin-signup">
+
+                <form v-on:submit.prevent="processSignUp" class="sign-in-form">
+
+                    <h2 class="title">Registrarse</h2>
+
+                    <div class="input-field">
+                        <i class="fas fa-envelope"></i>
+                        <input v-model="signupData.username" type="text" placeholder="Usuario"> 
+                    </div>
+                    <div class="input-field">
+                        <i class="fas fa-lock"></i>
+                        <input v-model="signupData.password1" type="password" placeholder="Contraseña">
+                    </div>
+                    <div class="input-field">
+                        <i class="fas fa-lock"></i>
+                        <input v-model="signupData.password2" type="password" placeholder="Verificar Contraseña">
+                    </div>
+                    <!-- <div class="input-field">
+                        <i class="fas fa-envelope"></i>
+                        <input v-model="signupData.email" type="text" placeholder="Correo"> 
+                    </div> -->
+                    <div class="input-field">
+                        <i class="fas fa-user"></i>
+                        <input v-model="signupData.nombre" type="text" placeholder="Nombre Completo"> 
+                    </div>
+                    <div class="input-field">
+                        <i class="fas fa-map-marker-alt"></i>
+                        <input v-model="signupData.direccion" type="text" placeholder="Dirección">
+                    </div>
+                    <div v-if="is_proveedor" class="input-field">
+                        <i class="fas fa-envelope"></i>
+                        <input v-model="signupData.nit" type="text" placeholder="NIT"> 
+                    </div>
+
+                    <button v-bind:class="{'disabled': is_loading}" class="btn solid">
+                        <span v-if="!is_loading">Registrarse</span>
+                        <div v-if="is_loading" class="spinner-border text-light" role="status"></div>
+                    </button>
+                </form>
+
+            </div>
+
+        </div>
+
+        <div class="panels-container">
+
+            <div class="panel left-panel">
+
+                <div class="content">
+
+                    <h3>¿Ya tienes una cuenta?</h3>
+                    <p>Puedes acceder para que sigas disfrutando de tu aplicación.</p>
+                    <button v-on:click="loadLogIn" class="btn transparent" id="sign-in-btn">Ingresar</button>
+                    
+                </div>
+                <div class="content">
+
+                    <h3>¡Quiero ser!</h3>
+                    <p>Elige si deseas contribuir como proveedor o disfrutar como cliente</p>
+                    <button v-if="!is_proveedor" v-on:click="setProveedor" class="btn transparent" id="sign-in-btn">
+                        <span v-if="!is_loading">Cliente</span>
+                    </button>
+                    <button v-if="is_proveedor" v-on:click="setProveedor" class="btn transparent" id="sign-in-btn">
+                        <span v-if="!is_loading">Proveedor</span>
+                    </button>
+                    
+                </div>
+                <img src="..\assets\registerUser.svg" alt="Register User" class="image">
+
+            </div>
+
+        </div>
+
+    </div>
+</template>
+
+<script>
+import gql from "graphql-tag";
+
+export default {
+    name: "SignUp", 
+    data: function() {
+        return {
+            show_error: false,
+            is_loading: false,
+            is_proveedor:false,
+            signupData:{
+                username: "",
+                password1: "",
+                password2: "",
+                nombre: "",
+                direccion: "",
+                nit: ""
+            }
+        }
+    }, 
+    methods: {
+        processSignUp: async function(){
+            this.is_loading = true;
+            this.signupData.nit = +this.signupData.nit;
+            if(this.is_proveedor){
+                await this.$apollo.mutate({
+                    mutation: gql`
+                    mutation SignUpProveedor($signupData: SignupInputProveedor!) {
+                        signUpProveedor(signupData: $signupData) {
+                            key
+                        }
+                    }`,
+                    variables: {
+                        signupData: this.signupData
+                    }
+                })
+                .then((result)=>{
+                    console.log("FUNCIONÓOOO")
+                    this.is_loading = false;
+                    this.show_error = false;
+                    let data = {
+                        username: this.signupData.username,
+                        token: result.data.signUpProveedor.key
+                    }
+                    this.$emit("completedLogin", data);
+                })
+                .catch((error)=>{
+                    this.show_error = true;
+                    console.log("DIO ERROR :c")
+                    console.log(error)
+                    this.is_loading = false;
+                })
+            } else {
+                await this.$apollo.mutate({
+                    mutation: gql`
+                    mutation SignUpCliente($signupData: SignupInputCliente!) {
+                        signUpCliente(signupData: $signupData) {
+                            key
+                        }
+                    }`,
+                    variables: {
+                        signupData: {
+                            username: this.signupData.username,
+                            password1: this.signupData.password1,
+                            password2: this.signupData.password2,
+                            nombre: this.signupData.nombre,
+                            direccion: this.signupData.direccion
+                        }
+                    }
+                })
+                .then((result)=>{
+                    console.log("FUNCIONÓOOO")
+                    this.is_loading = false;
+                    this.show_error = false;
+                    let data = {
+                        username: this.signupData.username,
+                        token: result.data.signUpCliente.key
+                    }
+                    this.$emit("completedLogin", data);
+                })
+                .catch((error)=>{
+                    this.show_error = true;
+                    console.log("DIO ERROR :c")
+                    console.log(error)
+                    this.is_loading = false;
+                })
+            }
+        },
+        
+        loadLogIn(){
+            this.$router.push({name: "logIn"});
+        },
+
+        setProveedor(){
+            var state = this.is_proveedor;
+            if(state){
+                this.is_proveedor = false;
+            } else {
+                this.is_proveedor = true;
+            }
+        }
+    }, 
+    created: function () {}
+};
+</script>
+
+<style scoped>
+.container {
+  position: relative;
+  width: 100%;
+  background-color: #fff;
+  min-height: 100vh;
+  overflow: hidden;
+}
+
+.forms-container {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+  }
+  
+
+.signin-signup {
+
+    position: absolute;
+    top: 50%; left: 75%;
+    transform: translate(-50%, -50%);
+    width: 50%;
+    transition: 1s 0.7s ease-in-out;
+    display: grid;
+    grid-template-columns: 1fr;
+    z-index: 5;
+
+}
+
+form {
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+    padding: 0 5rem;
+    transition: all 1s 0.7s;
+    overflow: hidden;
+    grid-column: 1 / 2;
+    grid-row: 1 / 2;
+
+}
+
+form.sign-up-form {
+
+    opacity: 0;
+    z-index: 1;
+
+}
+
+form.sign-in-form {
+
+    z-index: 2;
+
+}
+
+.title {
+
+    font-size: 2.2rem;
+    color: #444;
+    margin-bottom: 10px;
+
+}
+
+.input-field {
+
+    position: relative;
+    max-width: 380px;
+    width: 100%;
+    background-color: #f0f0f0;
+    margin: 10px 0;
+    height: 55px;
+    border-radius: 55px;
+    grid-template-columns: 15% 85%;
+    padding: 0 .4rem;
+
+}
+
+.input-field i {
+
+    text-align: center;
+    line-height: 55px;
+    color: #acacac;
+    transition: 0.5s;
+    font-size: 1.1rem;
+    margin: 0 10px;
+
+}
+
+.input-field input {
+
+    background: none;
+    border: none; outline: none;
+    line-height: 1;
+    font-weight: 600;
+    font-size: 1.1rem;
+    color: #333;
+
+}
+
+.input-field input::placeholder {
+
+    color: #aaa;
+    font-weight: 500;
+
+
+}
+
+.social-text {
+
+    padding: 0.7rem 0;
+    font-size: 1rem;
+
+}
+
+.social-media {
+
+    display: flex;
+    justify-self: center;
+    transition: 0.3s;
+    transform: translateY(-0px);
+
+}
+
+.social-icon {
+
+    height: 46px; width: 46px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 0 .45rem;
+    color: #333;
+    border: 1px solid #333;
+    border-radius: 50%;
+    text-decoration: none;
+    font-size: 1.1rem;
+    transition: 0.3s;
+
+}
+
+.social-icon:hover {
+
+    color: #fff;
+    background-color: #333;    
+
+}
+
+.btn {
+
+    width: 150px;
+    background-color: #3e6553;
+    color: #fff;
+    border: none; outline: none;
+    height: 49px;
+    border-radius: 49px;
+    font-size: 1rem;
+    text-transform: uppercase;
+    font-weight: 600;
+    margin: 10px 0;
+    cursor: pointer;
+    transition: 0.5s;
+
+}
+
+.btn:hover {
+
+    background-color: #365949;
+
+}
+
+.panels-container {
+
+    position: absolute;
+    top: 0; left: 0;
+    width: 100%; height: 100%;
+    background-color: #fff;
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+
+}
+
+.container:before {
+
+    content: '';
+    position: absolute;
+    top: -10px; right: 48%;
+    width: 2000px; height: 2000px;
+    transform: translateY(-50%);
+    background-image: linear-gradient(-45deg, #365949 0%, #93bdaa 100%);
+    transition: 1.8s ease-in-out;
+    border-radius: 50%;
+    z-index: 6;
+
+}
+
+.image {
+
+    width: 70%;
+    transition: transform 0.9s ease-in-out;
+    transition-duration: .6s;
+
+}
+
+.panel {
+
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    justify-content: space-between;
+    text-align: center;
+    z-index: 6;
+
+}
+
+.left-panel {
+
+    pointer-events: all;
+    padding: 3rem 17% 2rem 12%;
+
+}
+.right-panel {
+
+    pointer-events: none;
+    padding: 3rem 17% 2rem 17%;
+
+}
+
+.panel .content {
+
+    color: #fff;
+    transition: transform 0.9s ease-in-out;
+    transition-delay: .6s;
+
+}
+
+.panel h3 {
+
+    font-weight: 600;
+    line-height: 1;
+    font-size: 1.5rem;
+
+}
+
+.panel p {
+
+    font-size: .95rem;
+    padding: 0.7rem 0;
+
+}
+
+.btn.transparent {
+
+    margin: 0;
+    margin-top: .2rem;
+    background: none;
+    border: 2px solid #fff;
+    width: 130px; height: 41px;
+    font-weight: 600;
+    font-size: .8rem;
+
+}
+
+.error{
+    color: red;
+    font-size: 15px;
+}
+
+.disabled{
+    pointer-events: none;
+    opacity: 0.6;
+}
+
+.right-panel .image,
+.right-panel .content {
+
+    transform: translateX(800px);
+
+}
+
+/* ANIMACIONES */
+
+.container.sign-up-mode:before {
+    transform: translate(100%, -50%);
+    right: 52%;
+  }
+  
+  .container.sign-up-mode .left-panel .image,
+  .container.sign-up-mode .left-panel .content {
+    transform: translateX(-800px);
+  }
+  
+  .container.sign-up-mode .signin-signup {
+    left: 25%;
+  }
+  
+  .container.sign-up-mode form.sign-up-form {
+    opacity: 1;
+    z-index: 2;
+  }
+  
+  .container.sign-up-mode form.sign-in-form {
+    opacity: 0;
+    z-index: 1;
+  }
+  
+  .container.sign-up-mode .right-panel .image,
+  .container.sign-up-mode .right-panel .content {
+    transform: translateX(0%);
+  }
+  
+  .container.sign-up-mode .left-panel {
+    pointer-events: none;
+  }
+  
+  .container.sign-up-mode .right-panel {
+    pointer-events: all;
+  }
+
+  @media (max-width: 570px) {
+    form {
+      padding: 0 1.5rem;
+    }
+  
+    .image {
+      display: none;
+    }
+    .panel .content {
+      padding: 0.5rem 1rem;
+    }
+    .container {
+      padding: 1.5rem;
+    }
+  
+    .container:before {
+      bottom: 72%;
+      left: 50%;
+    }
+  
+    .container.sign-up-mode:before {
+      bottom: 28%;
+      left: 50%;
+    }
+  }
+
+  @media (max-width: 870px) {
+    .container {
+      min-height: 800px;
+      height: 100vh;
+    }
+    .signin-signup {
+      width: 100%;
+      top: 95%;
+      transform: translate(-50%, -100%);
+      transition: 1s 0.8s ease-in-out;
+    }
+  
+    .signin-signup,
+    .container.sign-up-mode .signin-signup {
+      left: 50%;
+    }
+  
+    .panels-container {
+      grid-template-columns: 1fr;
+      grid-template-rows: 1fr 2fr 1fr;
+    }
+  
+    .panel {
+      flex-direction: row;
+      justify-content: space-around;
+      align-items: center;
+      padding: 2.5rem 8%;
+      grid-column: 1 / 2;
+    }
+  
+    .right-panel {
+      grid-row: 3 / 4;
+    }
+  
+    .left-panel {
+      grid-row: 1 / 2;
+    }
+  
+    .image {
+      width: 200px;
+      transition: transform 0.9s ease-in-out;
+      transition-delay: 0.6s;
+    }
+  
+    .panel .content {
+      padding-right: 15%;
+      transition: transform 0.9s ease-in-out;
+      transition-delay: 0.8s;
+    }
+  
+    .panel h3 {
+      font-size: 1.2rem;
+    }
+  
+    .panel p {
+      font-size: 0.7rem;
+      padding: 0.5rem 0;
+    }
+  
+    .btn.transparent {
+      width: 110px;
+      height: 35px;
+      font-size: 0.7rem;
+    }
+  
+    .container:before {
+      width: 1500px;
+      height: 1500px;
+      transform: translateX(-50%);
+      left: 30%;
+      bottom: 68%;
+      right: initial;
+      top: initial;
+      transition: 2s ease-in-out;
+    }
+  
+    .container.sign-up-mode:before {
+      transform: translate(-50%, 100%);
+      bottom: 32%;
+      right: initial;
+    }
+  
+    .container.sign-up-mode .left-panel .image,
+    .container.sign-up-mode .left-panel .content {
+      transform: translateY(-300px);
+    }
+  
+    .container.sign-up-mode .right-panel .image,
+    .container.sign-up-mode .right-panel .content {
+      transform: translateY(0px);
+    }
+  
+    .right-panel .image,
+    .right-panel .content {
+      transform: translateY(300px);
+    }
+  
+    .container.sign-up-mode .signin-signup {
+      top: 5%;
+      transform: translate(-50%, 0);
+    }
+  }
+</style>
