+++ leaf-fe-vue/src/router.js
@@ -1,8 +1,9 @@
 import { createRouter, createWebHashHistory } from 'vue-router'
 import App from './App.vue'
-import LogIn from './components/login_signup.vue'
+import LogIn from './components/login.vue'
+import signUp from './components/signup.vue'
 import Carrito from './components/carrito.vue'
-/*import Index from './components/index.vue'*/
+import Index from './components/index.vue'
 const routes = [
   {
     path: '/',
@@ -15,15 +16,20 @@ const routes = [
     component: LogIn
   },
   {
+    path: '/signUp',
+    name: 'signUp',
+    component: signUp
+  },
+  {
     path: '/carrito',
     name: 'carrito',
     component: Carrito
   }, 
-  /*{
+  {
     path: '/index',
     name: 'index',
     component: Index
-  },*/
+  },
   /*{
     path: '/',
     name: 'recuperacion',
