+++ leaf-fe-vue/vue.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    devServer: {
+        disableHostCheck: true
+    }
+}
